refactor: move recovery migration to 8.3
The column family prefix correction migration task is introduced on
migration to 8.4, but has now been backported to 8.3.

On 8.3, we should therefore have this in the to_8_3 package.

I'm not sure whether this has any further effects, but it feels like it
makes sense to me.
