+++ engine/src/main/java/io/camunda/zeebe/engine/state/migration/to_8_3/ColumnFamilyCorrectionException.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.state.migration.to_8_4.corrections;
+package io.camunda.zeebe.engine.state.migration.to_8_3;
 
 import io.camunda.zeebe.db.impl.DbBytes;
 import io.camunda.zeebe.protocol.ZbColumnFamilies;
