+++ engine/src/main/java/io/camunda/zeebe/engine/state/migration/to_8_3/DbColumnFamilyCorrectionMigrationState.java
@@ -5,13 +5,10 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.state.migration.to_8_4;
+package io.camunda.zeebe.engine.state.migration.to_8_3;
 
 import io.camunda.zeebe.db.TransactionContext;
 import io.camunda.zeebe.db.ZeebeDb;
-import io.camunda.zeebe.engine.state.migration.to_8_4.corrections.ColumnFamily48Corrector;
-import io.camunda.zeebe.engine.state.migration.to_8_4.corrections.ColumnFamily49Corrector;
-import io.camunda.zeebe.engine.state.migration.to_8_4.corrections.ColumnFamily50Corrector;
 import io.camunda.zeebe.protocol.ZbColumnFamilies;
 
 public final class DbColumnFamilyCorrectionMigrationState {
