+++ engine/src/test/java/io/camunda/zeebe/engine/state/migration/to_8_3/ColumnFamilyPrefixCorrectionMigrationTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.state.migration.to_8_4;
+package io.camunda.zeebe.engine.state.migration.to_8_3;
 
 import io.camunda.zeebe.db.ColumnFamily;
 import io.camunda.zeebe.db.TransactionContext;
@@ -18,9 +18,6 @@ import io.camunda.zeebe.db.impl.DbString;
 import io.camunda.zeebe.engine.state.message.DbMessageState;
 import io.camunda.zeebe.engine.state.migration.MigrationTaskState;
 import io.camunda.zeebe.engine.state.migration.MigrationTaskState.State;
-import io.camunda.zeebe.engine.state.migration.to_8_4.corrections.ColumnFamily48Corrector;
-import io.camunda.zeebe.engine.state.migration.to_8_4.corrections.ColumnFamily49Corrector;
-import io.camunda.zeebe.engine.state.migration.to_8_4.corrections.ColumnFamily50Corrector;
 import io.camunda.zeebe.engine.state.mutable.MutableProcessingState;
 import io.camunda.zeebe.engine.state.signal.SignalSubscription;
 import io.camunda.zeebe.engine.util.ProcessingStateExtension;
